backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
litellm>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
